+++ slik/utils.py
@@ -1,4 +1,14 @@
 import pickle,yaml,os,pathlib,sys
+from sklearn.metrics import (accuracy_score,
+                             precision_score,
+                             recall_score,
+                             f1_score,
+                             confusion_matrix,
+                             roc_curve,
+                             roc_auc_score,
+                             precision_recall_curve,
+                             average_precision_score)
+                             
 
 def print_devider(title):
     print('\n{} {} {}\n'.format('-' * 25, title, '-' * 25))
@@ -16,7 +26,15 @@ def store_attribute(dict_file,output_path):
 def store_pipeline(pipeline_object, pipeline_path):
     # save the model to disk
     pickle.dump(pipeline_object, open(pipeline_path, 'wb'))
+
+def get_scores(y_true, y_pred):
     
+    return {
+      'accuracy': accuracy_score(y_true, y_pred),
+      'precision': precision_score(y_true, y_pred),
+      'recall': recall_score(y_true, y_pred),
+      'f1': f1_score(y_true, y_pred),
+    }
     
 class HiddenPrints:
     def __enter__(self):
